day 5: complete
what a pain... Just when I was going to go to bed the inverse solution yield the result.
A really shitty solution. The brute force was supposed to take 6 hours or so...
The only clever solution I could think of was to do and inverse traslation from location to soil
So I can go up until a find a seed in the ranges
I gave up on that solution yesterday, but in the end only needed about an hour or so
Happy this problem is over.