day 14: I hate loops
Loops are the bane of my existence. It was the same problem with tetris last year. I found the loop, easy. But it took me 2 days how
to translate those loops into a function able to calculate the status
at any point in time.

Also my tilt function is a mess, I could not get bothered trying to generalise for each direction. I did a shameful copy/pasta

Also I have a little bit of rant about dicts and hashkeys... Apparently
keys need to be a string.