day 17: dijkstra
Yep. It was about time to use our old friend. I liked the implementation  I used (dijkstar). Nice, clean and simple...

But it was not as straightforward to build the graph. My trick is to have 2 nodes per square (the horizontal and the vertical) and a cost function that penalises horizontal to horizonal or vertical to vertical edges

I'm suprised it worked