day 21: part 1
BFS. I I finally learned how to do variable type hinting. I struggle too much and in the end it was about odd/even (I had to use hints for that)

Part 2 is just unphantomable. No way in hell even trying the coordinate reduction I can get that number on the infinite map. There are hints about a quadratic equation but or about the rhombus pattern the input has but I think I'm going to park it for nnow