day 22: 3D block
It took much more than it should had... For once, I took advantage of python easy 3D rendering and I was able to plot the 3D grid which make debugging much easier.

Part 2... It took me some time to read it correctly. And it's really slow... but it does it in a reasonable time. I'm starting to remember how deques and BFS are much better than recursion