day 24: part 1 only
I did not have time for part 2. Busy day. My geometry it's a little bit rusty but I managed to figure out this was about interpolating trajectories. After a little bit of googling and refreshing my Maths for the intersect function, it was simple from there

(update for day 23... part 2 yield a result a little bit after I pushed the changes. It dit not complete but it was the right solution) I will probably revisit this to get something better but it's parked for now