day 12. Part 2
After days trying to figure out new algorithms, that in theory it will be non-recursive and cacheable... They did not work.

Looking for hints, it seems that my idea to try all the possible options was correct but with a couple of tweaks.

No more trying to match as much as possible, one group match until the end is reached. At the end if all the groups are consumed, it's a single match
The cache key it's the group index, not the matched groups. That means that the caches prunes a lot of the options already calculated.