day 24: Part 2
My Maths are not good enough for this... Another one I needed some hints. The problem is just to find 3 moments in time in which  the 3 hailstones are in the same line. That should be enough to calculate the velocity and the starting point.

That's a system with 9 equations and 9 variables ( starting position, speed and 3 moments in time)

Not even at my Math's prime (year one at Uni) I would had been able to solve that... So I looked for hints, to do Gauss Reduction (something that I may or may not have  studied in Algebra 202) I needed a crush course and some people mentioned z3. OMG, that library is pure magic.
I have to say that the python binding (with operator overloading) makes the thing beautifull.

This is something to add to my AoC list of tools for next year