day 21: Part 2
There is no way in hell I could have come with this solution. https://github.com/villuna/aoc23/wiki/A-Geometric-solution-to-advent-of-code-2023,-day-21

Other people used a quadratic ecuation interpolation... Far avoe my Maths level.

Quite disappointed this year. I failed on 3 days.

Python is ok-ish. It's still a bad language full of questionable choices (white spaces instead of block markers, no switch, no interfaces, right to left expresions instead of left to right.

yes, comprehensions are powerful, but theyr'e quite unreadable. At the end of the day python is as much unreadable as c. Yes it uses and and ors instead of symbols but there is so much magic on comprehensions and sublists that invites you to do one-lines.

One-lines are unreadble.

Type-hinting is even worse shoe-horned that in PHP

But now at least I'm fluent enoguh to do things with it in a reasoable time